// src/state/privilege/privilege-type.ts
export enum PrivilegeType {
  READ = 'READ',
  WRITE = 'WRITE',
  DASHBOARD_ADMIN = 'DASHBOARD_ADMIN',
  DASHBOARD_USER = 'DASHBOARD_USER',
  APP_SOURCE_ADMIN = 'APP_SOURCE_ADMIN',
  SCRATCH_READ = 'SCRATCH_READ',
  SCRATCH_WRITE = 'SCRATCH_WRITE',
  SCRATCH_ADMIN = 'SCRATCH_ADMIN',
  APP_CLIENT_DEVELOPER = 'APP_CLIENT_DEVELOPER',
  PERSON_CREATE = 'PERSON_CREATE',
  PERSON_EDIT = 'PERSON_EDIT',
  PERSON_DELETE = 'PERSON_DELETE',
  PERSON_READ = 'PERSON_READ',
  ORGANIZATION_CREATE = 'ORGANIZATION_CREATE',
  ORGANIZATION_EDIT = 'ORGANIZATION_EDIT',
  ORGANIZATION_DELETE = 'ORGANIZATION_DELETE',
  ORGANIZATION_READ = 'ORGANIZATION_READ',
}